Restructure the folder
