--- src/components/Dragon.js
@@ -1,20 +1,0 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-
-const Dragon = () => {
-  const dispatch = useDispatch();
-  const dragon = useSelector((state) => state.dragonReducer);
-  useEffect(() => {
-    dispatch(dragon());
-  }, []);
-
-  return (
-    <div>
-      <h1>My dragons</h1>
-      <h1>My dragons</h1>
-      <h1>My dragons</h1>
-    </div>
-  );
-};
-
-export default Dragon;
