+++ src/components/dragon/dragonContainer.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../header';
+import Header from '../header/NavBar';
 import Dragons from './dragons';
 
 const dragonContainer = () => (
