--- src/components/header/index.jsx
@@ -1,11 +1,0 @@
-/* eslint-disable jsx-quotes */
-import React from 'react';
-import '../../assets/css/style.css';
-
-const index = () => (
-  <div className='header'>
-    <h1>my header</h1>
-  </div>
-);
-
-export default index;
