+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import App from './App';
 import store from './redux/store';
-import './assets/css/main.css';
+import './assets/styles/main.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
